refactor welcomeview
